fixed bugs due to the previous merge
